+++ src/modules/llms/transports/server/anthropic.router.ts
@@ -4,8 +4,8 @@ import { TRPCError } from '@trpc/server';
 import { createTRPCRouter, publicProcedure } from '~/modules/trpc/trpc.server';
 import { fetchJsonOrTRPCError } from '~/modules/trpc/trpc.serverutils';
 
-import { chatGenerateOutputSchema, historySchema, modelSchema } from '../openai/openai.router';
-import { listModelsOutputSchema, LLM_IF_OAI_Chat, ModelDescriptionSchema } from '../llm.router';
+import { chatGenerateOutputSchema, historySchema, modelSchema } from './openai.router';
+import { listModelsOutputSchema, LLM_IF_OAI_Chat, ModelDescriptionSchema } from '../../llm.router';
 
 import { AnthropicWire } from './anthropic.wiretypes';
