+++ src/modules/llms/transports/server/openai.wiretypes.ts
