+++ src/modules/trpc/trpc.router.ts
@@ -2,9 +2,9 @@ import { createTRPCRouter } from './trpc.server';
 
 import { elevenlabsRouter } from '~/modules/elevenlabs/elevenlabs.router';
 import { googleSearchRouter } from '~/modules/google/search.router';
-import { llmAnthropicRouter } from '~/modules/llms/anthropic/anthropic.router';
-import { llmAzureRouter } from '~/modules/llms/azure/azure.router';
-import { llmOpenAIRouter } from '~/modules/llms/openai/openai.router';
+import { llmAnthropicRouter } from '~/modules/llms/transports/server/anthropic.router';
+import { llmAzureRouter } from '~/modules/llms/transports/server/azure.router';
+import { llmOpenAIRouter } from '~/modules/llms/transports/server/openai.router';
 import { prodiaRouter } from '~/modules/prodia/prodia.router';
 import { sharingRouter } from '~/modules/sharing/sharing.router';
 import { ytPersonaRouter } from '../../apps/personas/ytpersona.router';
